+++ src/main/java/ladder/domain/LadderFactory.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
