+++ src/test/java/ladder/domain/LadderTest.java
@@ -1,10 +1,13 @@
-package ladder;
+package ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.Collections;
 import java.util.List;
+import ladder.domain.Ladder;
+import ladder.domain.Line;
+import ladder.domain.LineFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
